requirements.txt
Flask
Flask-WTF
matplotlib
python-dotenv